refactor: move utils into shared
