+++ packages/atag/src/components/audio/index.html
@@ -96,7 +96,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 
 export default class Audio extends PolymerElement {
   static get is() {
