+++ packages/atag/src/components/canvas/index.html
@@ -6,10 +6,8 @@
 
 <script>
   import { PolymerElement } from '@polymer/polymer';
-  import {
-    afterNextRender,
-    throttle
-  } from '../../shared/utils';
+  import afterNextRender from '../../shared/afterNextRender';
+  import throttle from '../../shared/throttle';
 
   export default class CanvasElement extends PolymerElement {
     static get is() {
