+++ packages/atag/src/components/checkbox-group/index.html
@@ -16,7 +16,8 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender, throttle } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
+import throttle from '../../shared/throttle';
 
 export default class CheckboxGroupElement extends PolymerElement {
   static get is() {
