+++ packages/atag/src/components/checkbox/index.html
@@ -54,7 +54,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 
 export default class CheckboxElement extends PolymerElement {
   static get is() {
