+++ packages/atag/src/components/image/index.html
@@ -24,7 +24,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 
 function handleIntersect(entries) {
   entries.forEach(entry => {
