+++ packages/atag/src/components/input/index.html
@@ -29,7 +29,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 
 let uid = 0;
