+++ packages/atag/src/components/picker-view/index.html
@@ -15,7 +15,7 @@
 <script>
 import './picker-view-column';
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 
 export default class PickerView extends PolymerElement {
   static get is() {
