+++ packages/atag/src/components/picker-view/picker-view-column.html
@@ -58,7 +58,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 
 export default class PickerViewColumn extends PolymerElement {
   static get is() {
