+++ packages/atag/src/components/picker/index.html
@@ -139,10 +139,10 @@
 </dom-module>
 
 <script>
-import { PolymerElement } from '@polymer/polymer';
 import '@polymer/polymer/lib/elements/dom-repeat.js';
-
-import { throttle, easeInOutCubic } from '../../shared/utils';
+import { PolymerElement } from '@polymer/polymer';
+import throttle from '../../shared/throttle';
+import easeInOutCubic from '../../shared/easeInOutCubic';
 
 // After modifying this, the above CSS will also be modified.
 const LINEHEIGHT = 27 * 100 / document.documentElement.clientWidth;
