+++ packages/atag/src/components/radio-group/index.html
@@ -15,7 +15,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 
 export default class RadioGroupElement extends PolymerElement {
   static get is() {
