+++ packages/atag/src/components/radio/index.html
@@ -64,7 +64,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 
 export default class RadioElement extends PolymerElement {
   static get is() {
