+++ packages/atag/src/components/switch/index.html
@@ -49,7 +49,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils/';
+import afterNextRender from '../../shared/afterNextRender';
 
 export default class SwitchElement extends PolymerElement {
   static get is() {
