+++ packages/atag/src/components/textarea/index.html
@@ -48,7 +48,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 import autosize from './autosize';
 
 let uid = 0;
