+++ packages/atag/src/components/video/index.js
@@ -1,4 +1,4 @@
-import shouldDowngradeNativeView from '../../shared/utils/shouldDowngradeNativeView';
+import shouldDowngradeNativeView from '../../shared/shouldDowngradeNativeView';
 import NativeVideo from './video.native';
 import WebVideo from './video.web';
