+++ packages/atag/src/components/view/index.html
@@ -14,7 +14,7 @@
 
 <script>
 import { PolymerElement } from '@polymer/polymer';
-import { afterNextRender } from '../../shared/utils';
+import afterNextRender from '../../shared/afterNextRender';
 
 export default class ViewElement extends PolymerElement {
   static get is() {
