--- packages/atag/src/shared/utils/easeInOutCubic.js
@@ -1,3 +1,0 @@
-export default function easeInOutCubic(t) {
-  return t < .5 ? 4 * t * t * t : (t - 1) * (2 * t - 2) * (2 * t - 2) + 1;
-}
