--- packages/atag/src/shared/utils/shouldDowngradeNativeView.js
@@ -1,9 +1,0 @@
-const embedViewManager =
-  typeof __embed_view_manager__ !== "undefined" ? __embed_view_manager__ : null; // eslint-disable-line
-
-/**
- * check whether native view is supported
- */
-export default function shouldDowngradeNativeView() {
-  return embedViewManager ? embedViewManager.shouldDowngrade() : false;
-}
